List preservation order
These are notes to help understand removing duplicate items from a list without changing the order of the list. This is done using the list(dict.fromkeys(items)) method which basically converts the list into a dictionary creating key, value pairs and storing unique items in a new dictionary and removing any duplicates. Dictionary cannot have duplicate keys that's why it removes the duplicates. That is then converted back into a list using the list() built in function which we included as list(dict.fromkeys(items)